# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Helpy Api Documentaion",
        "description": "This documentation for Helpy",
        "contact": {
            "email": "abdulrahman.masoud.official@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get Porofile Data",
                "description": "This To Get Full Data About Auth User",
                "operationId": "getProfileData",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "Success Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/profile/update": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update Profile Data ",
                "description": "This To Update Full Data About Auth User",
                "operationId": "Update Profile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "f_name",
                                    "l_name",
                                    "email",
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "f_name": {
                                        "description": "You must add First Name",
                                        "type": "string"
                                    },
                                    "l_name": {
                                        "description": "You must add Last Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "You must add Email",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "You must add Phone Number",
                                        "type": "phone"
                                    },
                                    "avatar": {
                                        "description": "Avatar Is Optionel",
                                        "type": "file"
                                    },
                                    "password": {
                                        "description": "You must add Password",
                                        "type": "password"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "-",
                                            "female",
                                            "male"
                                        ]
                                    },
                                    "country": {
                                        "description": "Country Is Optional",
                                        "type": "string"
                                    },
                                    "city": {
                                        "description": "City Is Optional",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "Address Is Optional",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "Success Request",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        []
    ]
}